並列最貴會全部列出

+++ HomeWork_20230215_01/Product.cs
@@ -13,7 +13,7 @@ namespace HomeWork_20230215_01
         [Name("商品編號")] public string Id { get; set; }
         [Name("商品名稱")] public string Name { get; set; }
         [Name("商品數量")] public int Quantity { get; set; }
-        [Name("價格")] public int Price { get; set; }
+        [Name("價格")] public decimal Price { get; set; }
         [Name("商品類別")] public string Type { get; set; }
 
         public void setting(string i, string n, string q, string p, string t)
